requirements.txt
asgiref==3.6.0
boto3==1.26.80
botocore==1.29.80
Django==3.1.13
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-storages==1.10.1
djangorestframework==3.14.0
jmespath==1.0.1
jsonfield==3.1.0
mysqlclient==2.0.3
pymemcache==4.0.0
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.7.1
s3transfer==0.6.0
sh==1.11
six==1.10.0
sqlparse==0.4.3
swapper==1.3.0
urllib3==1.26.14
wrapt==1.10.5
